REFACTOR: moved StatisticsOverviewTable to outer class
